split out guid util function on its own
